More libraries: Foreign related modules.
Fix in uulib parsers (by Doaitse).

Some 64 bit platform C warning messages fixed.

Less strict recompilation check, because building libraries is taking some time now,
whilst not always necessary to compile everything, and becoming a bit of a hindrance whilst debugging the library.

+++ EHC/src/rts/prim.cc
@@ -383,7 +383,7 @@ PRIM GrWord primOpenFile(GrWord str, GrWord mode)
 	case 5: modestring = "r+"; break;
 	case 6: modestring = "wb"; break;
 	case 7: modestring = "w"; break;
-	default:  printf("primOpenFile: illegal mode %d\n", mode); fflush(stdout);
+	default:  printf("primOpenFile: illegal mode %d\n", (int)mode); fflush(stdout);
 	          return 0;	
 }
